% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbrat_SSD.R
\name{lbrat.SSD.All}
\alias{lbrat.SSD.All}
\title{LBRAT test or RGMMAT test using SSD format files}
\usage{
lbrat.SSD.All(SSD.INFO, lbrat_est.obj, ...)
}
\arguments{
\item{SSD.INFO}{SSD format information file, output of function "Open_SSD". The genome wide scan are run set by set.}

\item{lbrat_est.obj}{ouput from lbrat_est.R}

\item{...}{Other options of the LBRAT or RGMMAT test. Defined same as in function "lbrat_test()".}
}
\value{
reults of the LBRAT or RGMMAT test. First column contains batchID, second column contains SNP ID, third column concains prospective P-value and forth column contains retrospective P-value
}
\description{
LBRAT test or RGMMAT test using SSD format files
}
