% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbrat_est.R
\name{lbrat_est.gee}
\alias{lbrat_est.gee}
\title{GEE NULL model estimation}
\usage{
lbrat_est.gee(y.long, time, y.cov, timecov = TRUE, corstr = "ar1",
  tol = 10^-6, max.iter = 50)
}
\arguments{
\item{y.long}{Long-formatted phenotype vector}

\item{time}{Time covarites matched with phenotype vector}

\item{y.cov}{Covariate matrix denoting the covariate variables measured at each time}

\item{timecov}{Logical variable, indicating whether the time fixed effect is estimated}

\item{corstr}{String, correlation structure for GEE model, optional values are: 'ar1', 'ind', 'mixture'}

\item{tol}{Numeric, tolerance for the iterative estimation in when using mixture correlation structure}

\item{max.iter}{Numeric, the maximum count for the iterative estimation in when using mixture correlation structure}
}
\value{
This function returns a list object with model parameters and residuals of the NULL GEE model
}
\description{
This function estimate the parameters and residuals for the NULL model in LBRAT
}
