% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbrat_SSD.R
\name{lbrat_read_phe}
\alias{lbrat_read_phe}
\title{Read in Phenotype and Genotype Files}
\usage{
lbrat_read_phe(y.file, cov.file, plink.file)
}
\arguments{
\item{y.file}{Path to the input file}

\item{cov.file}{Path to the covariates file}

\item{plink.file}{Path to the plink file frefix}
}
\value{
A formatted dataframe used as L-BRAT/RGMMAT test input
}
\description{
This function loads a phenotype file, a covariates file and a plink fam file
it's able to match the phenotype ID with genotype ID and generate input dataframe
for L-BRAT/RGMMAT test.
}
