% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbrat_simu.R
\name{lbrat_simu}
\alias{lbrat_simu}
\title{Simulation for LBRAT test}
\usage{
lbrat_simu(n.sample = 1000, n.time = 5, par = list(),
  time_cov = TRUE, snp.count = 1000, intercept = TRUE,
  disease.para = list(), onlypower = FALSE, phe.model = "logistic",
  oversampling = "random")
}
\arguments{
\item{n.sample}{Numeric, sample size, number of individuals}

\item{n.time}{Numeric, number of measurements for each individual}

\item{par}{List, the parameters for the phenotype traits, including covaraites and individual specific time dependent random effects}

\item{time_cov}{Logical variable, indicating whether time effect is included in phenotypic traits}

\item{snp.count}{Numeric, number of SNPs}

\item{intercept}{Logical variable, indicating whether intercept is used in phenotypic traits}

\item{disease.para}{List, the parameters for disease allele and its effect size for power simulation}

\item{onlypower}{Logical variable, indicating whether include disease SNPs in the generated SNPs}

\item{phe.model}{String, the phenotype model, two optional values: 'logistic', 'liability'}

\item{oversampling}{String, the ascertainment scheme, three optional value: 'random', 'baseline', 'sum'}
}
\value{
A list object is returned to be used as object for LBRAT test
}
\description{
This function use pre-defined parameters to make the simulation data for the LBRAT test (including type I and power test)
}
