% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbrat_pvalue.R
\name{lbrat_test}
\alias{lbrat_test}
\title{Calculate prospective and retrospective P-values for GEE or GLMM model}
\usage{
lbrat_test(lbrat.est, G, impute.method = "fixed", GRM = NULL)
}
\arguments{
\item{lbrat.est}{The output of function "lbrat_est()"}

\item{G}{The genotype matrix, an m*q matrix where m is the number of subjects and q is the total number genetic variants.}

\item{impute.method}{choose the iputation method when there is missing genotype. Optional options are: 'random', 'fixed' or 'bestguess'.}

\item{GRM}{takes m-by-m genetic correlation matrix or kinship matrix.}
}
\value{
This function returns a dataframe. The row name is the SNP ID, the first column is the prospective score statistics, the second colum is the retrospective score statistics, the third column is the prospective pvalue and the forth column is the restrospective pvalue
}
\description{
This function tests a SNPs for a given SNP set for a given lbrat estimated null model.
}
